:radioactive: nuke service workers with three seperate methods for complete and utter annihilation
1. add a self destructing service worker to the site, when updated by the browser will disintegrate it self

This is the main method of getting rid of the service worker. According to spec, the browser will make sure to try to update the service worker after every 24 hours. It does so by hitting the installed service worker url, which for gatsby happens to be `/sw.js`. We just have to make sure our killswitch lies there and waits.

https://medium.com/@nekrtemplar/self-destroying-serviceworker-73d62921d717
https://github.com/NekR/self-destroying-sw/blob/master/packages/gatsby-plugin-remove-serviceworker/sw.js
https://love2dev.com/blog/how-to-uninstall-a-service-worker/

2. add javascript code manually unregister the service worker

if the update does not happen, but we manage get the new html response into the browser, this will clean up the service worker.

https://stackoverflow.com/a/33705250

3. add `clear-site-data: "storage"` header to tell the browser to kill the service worker

the intent here is to remove the service worker as fast as possible, even if no html get's through, if at least one response reaches the browser it will instantiate cleanup mode and kill it of.

unfortunatly safari does not support this header, which also means iPhone does not support it, even running chrome/firefox on the device, since apple requires custom browsers to use the safari engine undearneath.

accordning to Google Analytics 35% of our total users using safari,  and 70% of mobile users are using an iPhone.

so this is a nice bonus for the 65% users not on safari, and the remaining 35% will have to get their antibiotics pill from the other two methods.

https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Clear-Site-Data