:chart_with_upwards_trend: add Google Tag Manager
Was a bit confusing going into to this, so much termonologi and stuff
analytics, gtag, gtm.

As I understand it, GTM, or Google Tag Manager is a tracking orchestrator. Kind of like Kubernetes is to Docker.
But just adding the gtm once in our code, everything else can be controlled from the GTM dashboard, including Google Analytics.

Easy Peasy.

https://www.analyticsmania.com/post/gtag-vs-google-tag-manager/